pyproject.toml
[project]
name = "fantasy-movie-data-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.5.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "seaborn>=0.13.2",
    "streamlit>=1.56.0",
]
